// src/app/model/empleado.interface.ts
export interface Empleado {
  idemployee: number;
  paternal: String;
  maternal: String;
  names: String;
  login: String;
  pass: String;
  weekly_hours: String;
  extra_hours: number;
  extra_minutes: number;
  gender: String;
  dni: String;
  mobile: String;
}
/*"idemployee": 0,
"paternal": "paterno",
"maternal": "materno",
"names": "nombres",
"login": "46543318",
"pass": "",
"weekly_hours": "24",
"extra_hours": 0,
"extra_minutes": 0,
"gender": "M",
"dni": "88888888",
"mobile": "999999999"*/